--- agent_runtimes/chat/__init__.py
@@ -1,5 +1,0 @@
-# Copyright (c) 2025-2026 Datalayer, Inc.
-# Distributed under the terms of the Modified BSD License.
-# Copyright (c) 2023-2024 Datalayer, Inc.
-#
-# BSD 3-Clause License
