+++ agent_runtimes/config/__init__.py
@@ -8,7 +8,7 @@ Provides frontend configuration services that can be used by both
 Jupyter and FastAPI servers.
 """
 
-from .frontend import get_frontend_config
+from .frontend_config import get_frontend_config
 
 __all__ = [
     "get_frontend_config",
