--- agent_runtimes/jupyter/handlers/chat/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2025-2026 Datalayer, Inc.
-# Distributed under the terms of the Modified BSD License.
-
-"""Request handlers for the Agent Runtimes application."""
