+++ package.json
@@ -118,9 +118,9 @@
     "@ag-ui/encoder": "^0.0.42",
     "@ag-ui/proto": "^0.0.42",
     "@agentclientprotocol/sdk": "^0.8.0",
-    "@ai-sdk/react": "^3.0.5",
+    "@ai-sdk/react": "^3.0.29",
     "@anthropic-ai/sdk": "^0.52.0",
-    "@datalayer/core": "^0.0.21",
+    "@datalayer/core": "^0.0.23",
     "@datalayer/icons-react": "^1.0.6",
     "@datalayer/jupyter-lexical": "^1.0.6",
     "@datalayer/jupyter-react": "^2.0.0",
